+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased][]
 
+### Changed
+
+-   Use `enhanced-resolve` resolver for `postcss-import`
+
 ### Fixed
 
 -   Resolving paths when source is `data`
@@ -12,6 +16,4 @@
 
 -   Initial implementation
 
-[unreleased]:
-	https://github.com/niksy/sass-module-resolve-importer/compare/v1.0.0...HEAD
 [1.0.0]: https://github.com/niksy/sass-module-resolve-importer/tree/v1.0.0